#12328: Refactor into own header file
